requirements.txt
audioop-lts==0.2.1
pillow==11.3.0
PyAudio==0.2.14
pydub==0.25.1
